Consolidating the conversion of virtual aliases so they are consistent wherever they are used. Changes to: - Make aggregation clauses the same as virtual attributes in every way - Use virtual attributes in order by clauses as aliases - Add stirng functions to math and func nodes Kill uneeded terminating php tag Unit test Fixing test problem with aggregate clauses, that revealed additional problems around the issue of OnlyFullGroupBy databases.
Changed PostgresSql to be OnlyFullGroupBy by default.

Added a database adapter option so you can override the default when you set up the adapter.
Fixing error related to problem with precision in database declaration. MYSQL standard says specifying DECIMAL with no precision should result in zero decimal places. POSTGRESQL ignores the standard, and we had the unintended consequence of clobbering some of the test data in MYSQL. This makes both sides equal, and fixes the test to check for a decimal place.